requirements.txt
appdirs==1.4.4
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.14
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
coverage==6.4.4
cssselect==1.1.0
dateparser==1.1.2
demjson3==3.0.5
Django==3.2
django-rest-framework==0.1.0
djangorestframework==3.13.1
drf-yasg==1.21.3
facebook-scraper==0.2.59
facebook-sdk==3.1.0
fake-useragent==0.1.11
idna==3.4
importlib-metadata==5.0.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
packaging==21.3
parse==1.19.0
pyee==8.2.2
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==1.4.3
python-dateutil==2.8.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.3.2
requests==2.28.1
requests-html==0.10.0
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tqdm==4.64.1
tzdata==2022.5
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.12
w3lib==2.0.1
websockets==10.3
zipp==3.9.0
